+++ eyes.sdk.core/src/main/java/com/applitools/eyes/GetAccessibilityRegion.java
@@ -1,0 +1,9 @@
+package com.applitools.eyes;
+
+import com.applitools.eyes.fluent.GetRegion;
+
+import java.util.List;
+
+public interface GetAccessibilityRegion extends GetRegion {
+    List<AccessibilityRegionByRectangle> getRegions(EyesScreenshot screenshot);
+}