+++ eyes.sdk.core/src/main/java/com/applitools/eyes/IDriverProvider.java
@@ -1,5 +1,0 @@
-package com.applitools.eyes;
-
-public interface IDriverProvider {
-
-}