+++ eyes.sdk.core/src/main/java/com/applitools/eyes/IGetAccessibilityRegion.java
@@ -1,8 +1,0 @@
-package com.applitools.eyes;
-
-import java.util.List;
-
-public interface IGetAccessibilityRegion {
-    List<AccessibilityRegionByRectangle> getRegions(IDriverProvider eyesBase, EyesScreenshot screenshot);
-
-}