+++ eyes.sdk.core/src/main/java/com/applitools/eyes/fluent/GetFloatingRegion.java
@@ -1,11 +1,10 @@
 package com.applitools.eyes.fluent;
 
-import com.applitools.eyes.EyesBase;
 import com.applitools.eyes.EyesScreenshot;
 import com.applitools.eyes.FloatingMatchSettings;
 
 import java.util.List;
 
-public interface GetFloatingRegion {
-    List<FloatingMatchSettings> getRegions(EyesBase eyesBase, EyesScreenshot screenshot);
+public interface GetFloatingRegion extends GetRegion {
+    List<FloatingMatchSettings> getRegions(EyesScreenshot screenshot);
 }