+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/fluent/ISeleniumCheckTarget.java
@@ -5,7 +5,7 @@ import org.openqa.selenium.WebElement;
 
 import java.util.List;
 
-public interface ISeleniumCheckTarget extends IScrollRootElementContainer {
+public interface ISeleniumCheckTarget extends IScrollRootElementContainer, ImplicitInitiation {
     By getTargetSelector();
     WebElement getTargetElement();
     List<FrameLocator> getFrameChain();