+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/fluent/ImplicitInitiation.java
@@ -1,0 +1,8 @@
+package com.applitools.eyes.selenium.fluent;
+
+import com.applitools.eyes.Logger;
+import org.openqa.selenium.WebDriver;
+
+public interface ImplicitInitiation {
+    void init(Logger logger, WebDriver driver);
+}