+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/rendering/IGetSeleniumRegion.java
@@ -1,10 +1,9 @@
 package com.applitools.eyes.selenium.rendering;
 
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
 import java.util.List;
 
 public interface IGetSeleniumRegion {
-    List<WebElement> getElements(WebDriver webDriver);
+    List<WebElement> getElements();
 }