+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/model/RenderingInfo.java
@@ -6,12 +6,16 @@ public class RenderingInfo {
     private String accessToken = null;
     private String resultsUrl = null;
     private String stitchingServiceUrl = null;
+    private int maxImageHeight;
+    private int maxImageArea;
 
-    public RenderingInfo(String serviceUrl, String accessToken, String resultsUrl, String stitchingServiceUrl) {
+    public RenderingInfo(String serviceUrl, String accessToken, String resultsUrl, String stitchingServiceUrl, int maxImageHeight, int maxImageArea) {
         this.serviceUrl = serviceUrl;
         this.accessToken = accessToken;
         this.resultsUrl = resultsUrl;
         this.stitchingServiceUrl = stitchingServiceUrl;
+        this.maxImageHeight = maxImageHeight;
+        this.maxImageArea = maxImageArea;
     }
 
     public RenderingInfo() {
@@ -21,11 +25,9 @@ public class RenderingInfo {
         return serviceUrl;
     }
 
-
     public String getAccessToken() {
         return accessToken;
     }
-
 
     public String getResultsUrl() {
         return resultsUrl;
@@ -35,7 +37,11 @@ public class RenderingInfo {
         return stitchingServiceUrl;
     }
 
-    public void setStitchingServiceUrl(String stitchingServiceUrl) {
-        this.stitchingServiceUrl = stitchingServiceUrl;
+    public int getMaxImageHeight() {
+        return maxImageHeight;
+    }
+
+    public int getMaxImageArea() {
+        return maxImageArea;
     }
 }