+++ CHANGELOG.md
@@ -1,12 +1,12 @@
-## [vNext]
+## [3.170.0] - 2020-07-13
 ### Fixed
 - Fixed a bug when new tests weren't defined as new in old versions of the server. [Trello 1993](https://trello.com/c/JSnJauTu)
 ### Updated
 - Update all `GetRegion` classes, so they don't depend on `Eyes` or `EyesBase` objects. [Trello 1980](https://trello.com/c/RYAOPRpc)
-- Updated the render request to match the protocol. `sizeMode` was change to `target` and `stitchingServiceUrl` was changed to `stitchingService`. [Trello 1988](https://trello.com/c/Yr6EsUlL)
+- Updated the render request to match the protocol (internal). [Trello 1988](https://trello.com/c/Yr6EsUlL)
 - Limited screenshot size. [Trello 1991](https://trello.com/c/2iCNfoI7)
 
-## [3.169.0] - 2020-05-07
+## [3.169.0] - 2020-07-05
 ### Fixed
 - Correct size calculation for elements that change when hiding scrollbars. [Trello 1881](https://trello.com/c/9pVjoVwC) 
 ### Updated